requirements.txt
requests
python-dotenv
openai
pydantic

jupyter
ipykernel


instructor
anthropic
google-cloud-vision
fal-client

streamlit

fastapi
uvicorn



python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
fastapi-jwt
sqlalchemy
pyodbc